+++ conf/conf.go
@@ -66,6 +66,7 @@ type (
 		Topology        string       `yaml:"topology"`
 		InetPort        int          `yaml:"inet_port"`
 		Uid             *int         `yaml:"uid"`
+		ReadOnly        *bool        `yaml:"read_only"`
 		Dockerfile      string       `yaml:"dockerfile"`
 		DockerfileBuild string       `yaml:"dockerfile_build"`
 		HealthCheck     *HealthCheck `yaml:"health_check"`
@@ -225,6 +226,10 @@ func (recv *Config) SetDefaults() {
 
 				if region.Containers[0].Topology == "" {
 					region.Containers[0].Topology = Topology_Inet
+				}
+
+				if region.Containers[0].Name == "" {
+					region.Containers[0].Name = "primary"
 				}
 			}
 		}